aa/backend/responses/requirements.txt
Flask
python-dotenv
supabase
Flask-Cors
Flask-APScheduler
gunicorn 
requests 
